Create python_basic_reupload
Did you check these against the website at all? #1 {"error": "Error. Something went wrong trying to run your code. This may be a problem with the server and not with your code (e.g., someone may have forced the server to restart by writing an infinite loop). Please try again in a few seconds. If the error persists, please contact your instructors."} #2 {"error": "Error. Something went wrong trying to run your code. This may be a problem with the server and not with your code (e.g., someone may have forced the server to restart by writing an infinite loop). Please try again in a few seconds. If the error persists, please contact your instructors."} #3 {"error": "Error. Something went wrong trying to run your code. This may be a problem with the server and not with your code (e.g., someone may have forced the server to restart by writing an infinite loop). Please try again in a few seconds. If the error persists, please contact your instructors."} You need to return the final answer, not print it. We just use the website to check the code.